Use more auto, and c++ casting

+++ src/engine/Engine-main.cpp
@@ -76,7 +76,7 @@ void	CEngine::SetNodePosRot(GLuint id, glm::vec3 const & pos, glm::vec3 const & 
 	{
 		m_vObjectNode[id].position	=	pos;
 		for (unsigned int i=0; i < 3; ++i)
-			m_vObjectNode[id].pitch[i]	=	(float)pitch[i];
+			m_vObjectNode[id].pitch[i]	=	pitch[i];
 	}
 	// else we do nothing improve performance xD
 }
